Eliminated compilation warnings in wave.cpp.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@5075 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/sound/loader/wave.cpp
@@ -101,7 +101,7 @@ void csSoundDataWave::ConvertChannels(int Channels) {
   Type *NewData=new Type[NewNumSamples*Channels];               \
   Type *OldData=(Type*)Data;                                    \
   for (unsigned long i=0;i<NewNumSamples;i++) {                 \
-    int samppos = i/Factor;                                     \
+    int samppos = (int)(i/Factor);                              \
     if (Channels==1) {                                          \
       NewData[i]=OldData[samppos];                              \
     } else {                                                    \
@@ -117,7 +117,7 @@ void csSoundDataWave::ConvertChannels(int Channels) {
 void csSoundDataWave::ConvertFreq(int NewFreq) {
   if (NewFreq==Format.Freq) return;
   float Factor=NewFreq/Format.Freq;
-  unsigned long NewNumSamples=NumSamples*Factor;
+  unsigned long NewNumSamples=(unsigned long)(NumSamples*Factor);
   if (Format.Bits==16) {
     CONVERT_FREQ_TYPE(short,Format.Channels);
   } else {
